requirements.txt
bottle==0.12.25
python-dotenv==1.0.0
debugpy==1.8.0 